+++ Source/SSCore/src/main/java/com/realtech/socialsurvey/core/dao/impl/UserRankingThisMonthBranchDaoImpl.java
@@ -53,7 +53,7 @@ public class UserRankingThisMonthBranchDaoImpl extends GenericReportingDaoImpl<U
                 userRankingThisMonthBranch.setLastName( String.valueOf( row[3] ) );
                 userRankingThisMonthBranch.setRankingScore( Float.valueOf( String.valueOf( row[4] ) ) );
                 userRankingThisMonthBranch.setTotalReviews( Integer.valueOf( String.valueOf( row[5] ) ) );
-                userRankingThisMonthBranch.setAverageRating( Integer.valueOf( String.valueOf( row[6] ) )  );
+                userRankingThisMonthBranch.setAverageRating( Float.valueOf( String.valueOf( row[6] ) )  );
                 userRankingThisMonthBranch.setSps(  Float.valueOf( String.valueOf( row[7] ) )  );
                 userRankingThisMonthBranch.setCompletedPercentage(  Float.valueOf( String.valueOf( row[8] ) ) );
                 userRankingThisMonthBranch.setIsEligible( Integer.valueOf( String.valueOf( row[9] ) ) );
